Refactoring the Staff Model
