+++ src/main/java/seedu/inventory/model/StaffList.java
@@ -1,12 +1,12 @@
-package seedu.address.model;
+package seedu.inventory.model;
 
 import static java.util.Objects.requireNonNull;
 
 import java.util.List;
 
 import javafx.collections.ObservableList;
-import seedu.address.model.staff.Staff;
-import seedu.address.model.staff.UniqueStaffList;
+import seedu.inventory.model.staff.Staff;
+import seedu.inventory.model.staff.UniqueStaffList;
 
 /**
  * Wraps all data at the StaffList level
