+++ src/main/java/seedu/inventory/model/staff/exceptions/DuplicateStaffException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.staff.exceptions;
+package seedu.inventory.model.staff.exceptions;
 
 /**
  * Signals that the operation will result in duplicate Staff (Staffs are considered duplicates if they have the same
